COMPLETE: 9-Emotion Mood System + Remove Palette System
 FULLY OPERATIONAL - All systems working on ESP32-S3

Major Achievements:
- Added EXCITEMENT as 9th emotional state to mood system
- Successfully removed entire palette system, replaced with direct RGB565 colors
- Fixed ENABLE_GOBLIN_COMPONENTS build configuration issue
- All goblin components (left eye, right eye, nose) now functional
- Mood-driven animation engine working with automatic recomputation
- Build system stable: RAM 5.8%, Flash 51.7%
- Firmware uploaded and running successfully on COM4

Technical Changes:
- Updated p32_display.c: removed palette dependencies, added EXCITEMENT case
- Fixed p32_component_config.h: added missing ENABLE_GOBLIN_COMPONENTS define
- Updated component files: goblin_eye_left.c, goblin_eye_right.c, goblin_nose.c
- Created SESSION_CONTEXT_SUMMARY.md for future session startup efficiency

Serial Monitor Confirms:
- Components executing at loop 1900+ successfully
- Eye animations cycling (blink  angry stare) with dynamic RGB565 colors
- Sensor integration working (distance measurements)
- No palette-related errors, clean mood transitions

Next: Ready for full-body architecture expansion (arms/torso/legs)

+++ src/component_tables.c
@@ -11,11 +11,11 @@ extern esp_err_t goblin_eye_left_init(void);
 extern esp_err_t goblin_eye_right_init(void);
 extern esp_err_t goblin_nose_init(void);
 
-extern void system_act_component(uint32_t loopCount);
-extern void network_monitor_act(uint32_t loopCount);
-extern void goblin_eye_left_act(uint32_t loopCount);
-extern void goblin_eye_right_act(uint32_t loopCount);
-extern void goblin_nose_act(uint32_t loopCount);
+extern void system_act_component(uint64_t loopCount);
+extern void network_monitor_act(uint64_t loopCount);
+extern void goblin_eye_left_act(uint64_t loopCount);
+extern void goblin_eye_right_act(uint64_t loopCount);
+extern void goblin_nose_act(uint64_t loopCount);
 
 // Component initialization table - ACTUAL DEFINITION
 init_func_t initTable[INIT_TABLE_SIZE] = {
