+++ src/system_components.c
@@ -17,12 +17,12 @@ esp_err_t p32_comp_heartbeat_init(void) {
     return ESP_OK;
 }
 
-void p32_comp_heartbeat_act(uint32_t loopCount) {
+void p32_comp_heartbeat_act(uint64_t loopCount) {
 #ifdef SIMPLE_TEST
-    printf("ACT[%lu]: heartbeat - hitCount:1\n", loopCount);
+    printf("ACT[%llu]: heartbeat - hitCount:1\n", loopCount);
     return;
 #endif
-    ESP_LOGI(TAG_HEARTBEAT, "Loop %lu: System heartbeat", loopCount);
+    ESP_LOGI(TAG_HEARTBEAT, "Loop %llu: System heartbeat", loopCount);
 }
 
 esp_err_t p32_comp_network_monitor_init(void) {
@@ -34,18 +34,18 @@ esp_err_t p32_comp_network_monitor_init(void) {
     return ESP_OK;
 }
 
-void p32_comp_network_monitor_act(uint32_t loopCount) {
+void p32_comp_network_monitor_act(uint64_t loopCount) {
 #ifdef SIMPLE_TEST
     // Send timing packet to server every 100,000 loops for server-side timing measurement
     static uint32_t packet_counter = 0;
     if (loopCount % 100000 == 0 && loopCount > 0) {
-        printf("TIMING_PACKET: loop=%lu, packet=%lu\n", loopCount, packet_counter++);
+        printf("TIMING_PACKET: loop=%llu, packet=%lu\n", loopCount, packet_counter++);
         // TODO: Send actual network packet to server with loopCount and packet_counter
         // Server will measure time between packets to calculate loop performance
     }
-    if (loopCount % 1 == 0) printf("ACT[%lu]: network_monitor - hitCount:1\n", loopCount);
+    if (loopCount % 1 == 0) printf("ACT[%llu]: network_monitor - hitCount:1\n", loopCount);
     return;
 #endif
-    ESP_LOGI(TAG_NETWORK_MONITOR, "Loop %lu: Network monitoring and loop timing", loopCount);
+    ESP_LOGI(TAG_NETWORK_MONITOR, "Loop %llu: Network monitoring and loop timing", loopCount);
     // TODO: Implement actual network monitoring and server communication
 }
